+++ Unit 2/Assignment5/Assignment5/DetailViewController.m
@@ -63,7 +63,7 @@
     
 }
 
--(void) setSongIndex:(NSInteger) index
+-(void) setSongIndex:(long) index
 {
     _songIndex = index;
